timer minute and simulation service
