+++ SimulationService/src/main/java/org/energygrid/east/simulationservice/SimulationserviceApplication.java
@@ -1,0 +1,13 @@
+package org.energygrid.east.simulationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SimulationserviceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SimulationserviceApplication.class, args);
+    }
+
+}
