+++ SimulationService/src/main/java/org/energygrid/east/simulationservice/controller/SimulationController.java
@@ -1,24 +1,16 @@
-package org.energygrid.east.solarparkservice.controller;
+package org.energygrid.east.simulationservice.controller;
 
-import org.energygrid.east.solarparkservice.model.Simulation;
-import org.energygrid.east.solarparkservice.model.SimulationTimer;
-import org.energygrid.east.solarparkservice.model.SolarPark;
-import org.energygrid.east.solarparkservice.service.ISimulation;
-import org.energygrid.east.solarparkservice.service.ISolarParkPower;
-
+import org.energygrid.east.simulationservice.model.Simulation;
+import org.energygrid.east.simulationservice.service.ISimulation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.constraints.NotNull;
-import java.util.Timer;
 
 @RestController
 @RequestMapping("simulation")
 public class SimulationController {
-
-    @Autowired
-    private ISolarParkPower solarParkPowerService;
 
     @Autowired
     private ISimulation simulationService;
@@ -32,23 +24,14 @@ public class SimulationController {
 
     @GetMapping("/add")
     public ResponseEntity<Simulation> addSimulation() {
-        Timer timer = new Timer();
-
-        SolarPark solarPark = solarParkPowerService.getSolarParkByName("test");
-        Simulation simulation = new Simulation(solarPark, timer);
+        Simulation simulation = new Simulation();
         simulationService.addSimulation(simulation);
-
-        SimulationTimer simulationTimer = new SimulationTimer(simulation);
-        timer.scheduleAtFixedRate(simulationTimer, 1000, 1000);
-
         return ResponseEntity.status(200).body(simulation);
     }
 
     @DeleteMapping("/delete")
     public ResponseEntity<String> deleteSimulation(@RequestParam(name = "id") String id) {
-        Simulation simulation = simulationService.getSimulationById(id);
-        simulation.stopTimer();
-
+        simulationService.deleteSimulation(id);
         return ResponseEntity.status(200).body("Simulation: " + id + " stopped!");
     }
 }
