+++ SimulationService/src/main/java/org/energygrid/east/simulationservice/model/Simulation.java
@@ -1,0 +1,17 @@
+package org.energygrid.east.simulationservice.model;
+
+import java.util.Timer;
+import java.util.UUID;
+
+public class Simulation {
+
+    private final String id;
+
+    public Simulation() {
+        this.id = UUID.randomUUID().toString();
+    }
+
+    public String getId() {
+        return id;
+    }
+}
