+++ SimulationService/src/main/java/org/energygrid/east/simulationservice/service/ISimulation.java
@@ -1,0 +1,12 @@
+package org.energygrid.east.simulationservice.service;
+
+import org.energygrid.east.simulationservice.model.Simulation;
+
+public interface ISimulation {
+
+    Simulation getSimulationById(String id);
+
+    void addSimulation(Simulation simulation);
+
+    void deleteSimulation(String id);
+}
