+++ SimulationService/src/main/java/org/energygrid/east/simulationservice/service/SimulationService.java
@@ -1,13 +1,13 @@
-package org.energygrid.east.solarparkservice.service;
+package org.energygrid.east.simulationservice.service;
 
-import org.energygrid.east.solarparkservice.model.Simulation;
+import org.energygrid.east.simulationservice.model.Simulation;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
-public class SimulationService implements ISimulation {
+public class SimulationService implements ISimulation{
 
     private final List<Simulation> simulations;
 
@@ -23,5 +23,11 @@ public class SimulationService implements ISimulation {
     @Override
     public void addSimulation(Simulation simulation) {
         simulations.add(simulation);
+    }
+
+    @Override
+    public void deleteSimulation(String id) {
+        Simulation simulation = simulations.stream().filter(s -> s.getId().equals(id)).findFirst().orElse(null);
+        simulations.remove(simulation);
     }
 }
