+++ SimulationService/src/test/java/org/energygrid/east/simulationservice/SimulationserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package org.energygrid.east.simulationservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SimulationserviceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
