--- SolarParkService/src/main/java/org/energygrid/east/solarparkservice/service/ISimulation.java
@@ -1,10 +1,0 @@
-package org.energygrid.east.solarparkservice.service;
-
-import org.energygrid.east.solarparkservice.model.Simulation;
-
-public interface ISimulation {
-
-    Simulation getSimulationById(String id);
-
-    void addSimulation(Simulation simulation);
-}
