+++ pom.xml
@@ -47,5 +47,6 @@
         <module>gateway</module>
         <module>WeatherService</module>
         <module>AuthenticationService</module>
+        <module>SimulationService</module>
     </modules>
 </project>
